requirements.txt
attrdict==2.0.1
biome==0.1.3
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
idna==3.10
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathlib==1.0.1
pathspec==0.12.1
platformdirs==4.3.8
pyright==1.1.402
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
tk==0.1.0
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
